[level 1] Title: 부족한 금액 계산하기, Time: 0.12 ms, Memory: 85.2 MB -BaekjoonHub

+++ 프로그래머스/1/82612. 부족한 금액 계산하기/부족한 금액 계산하기.java
@@ -9,6 +9,6 @@ class Solution {
             num += price*i;
         }
         
-        return num<=money ? 0 : ((long)(num-money));
+        return num<=money ? 0 : num-money;
     }
 }
